+++ src/machine/constants.ts
@@ -1,1 +1,7 @@
 export const EVENTS_FULL = 'machine$$full';
+
+class Rinit {}
+
+export const RINIT_STATE = new Rinit();
+
+export type { Rinit };